requirements.txt
astro-sdk-python[snowflake,sqlite]
python-dotenv
pytest
pytest-dotenv
